\name{D3M-package}
\alias{D3M-package}
\alias{D3M}
\docType{package}
\title{
Two Sample Test with Wasserstein Metric
}
\description{
Two sample test based on Wasserstein metric. This is motivated from detection of differential DNA-methylation sites based on underlying distributions.
}
\details{
\tabular{ll}{
Package: \tab D3M\cr
Type: \tab Package\cr
Version: \tab 0.43\cr
Date: \tab 2015-09-27\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Yusuke Matsui & Teppei Shimamura
Maintainer: Yusuke Matsui <ymatsui@med.nagoya-u.ac.jp>
}
