% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_d3m.R
\name{d3m.plot}
\alias{d3m.plot}
\title{Comparative Plot of Two Distributions}
\usage{
d3m.plot(cases, control)
}
\arguments{
\item{cases}{name of case group data (vector)}

\item{control}{name of control group data (vector)}
}
\value{
beanplot and Q-Q plot of comparing distribution pairs.
}
\description{
Comparative Plot of Two Distributions
}
\examples{
library(D3M)
cases <- rbeta(50,1,5)
control <- rbeta(50,2,5)
d3m.plot(cases,control)
}
\author{
Yusuke Matsui & Teppei Shimamura
}

