% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wasser_vec.R
\name{wasserMetric.v}
\alias{wasserMetric.v}
\title{Wasserstein Metric}
\usage{
wasserMetric.v(cases, control, paranum = 101, q = 2)
}
\arguments{
\item{cases}{name of case group data (matrix sample * feature).}

\item{control}{names of control group data (matrix sample * feature).}

\item{paranum}{the number of quatile discretization + 1. Default is discretized by 1\%.}

\item{q}{power of Wasserstein metric. Default is q = 2.}
}
\value{
Wasserstein metric
}
\description{
Wasserstein Metric
}
\examples{
cases <- rbeta(30,1,5)
control <- rbeta(30,2,5)
wasserMetric.v(cases,control)
}
\author{
Yusuke Matsui & Teppei Shimamura
}

